refactor(slack-bot): decouple scrapers
