+++ lib/crawler/joke_image.rb
@@ -1,17 +1,13 @@
 require 'httparty'
 require 'nokogiri'
 
+require_relative './image_scraper'
+
 # Image based jokes implementation
 class ImageClone
-  def initialize
-    @base_url = 'https://www.hongkiat.com/blog/programming-jokes/'
-    unparsed_page = HTTParty.get(@base_url)
-    @parsed_page = Nokogiri::HTML(unparsed_page.body)
-  end
-
-  def image_jokes
+  def image_jokes(scrap)
     my_array = []
-    img_tags = @parsed_page.css('figure.entry-image img[src]').select { |image| image['src'].start_with?('https') }
+    img_tags = scrap.css('figure.entry-image img[src]').select { |image| image['src'].start_with?('https') }
     if img_tags.is_a? Array
       img_tags.map { |t| t[:src] }
     else
@@ -20,8 +16,9 @@ class ImageClone
   end
 
   def image_random_joke
-    if image_jokes.any?
-      image_jokes.sample
+    image = ImageScraper.new
+    if image_jokes(image.scraper).any?
+      image_jokes(image.scraper).sample
     else
       'Whoops, something is not quite right!, try again'
     end
@@ -30,3 +27,5 @@ class ImageClone
   private :image_jokes
 end
 
+image_clone = ImageClone.new
+puts image_clone.image_random_joke
